requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
data-processing==0.1.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
flatbuffers==24.3.25
gast==0.6.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
keras==3.6.0
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
selenium==4.25.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==8.3.2
tomli==2.0.2
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.16
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.6
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
